+++ source/client/eterGameLib/TerrainQuadtree.h
@@ -16,16 +16,16 @@ public:
 	virtual ~CTerrainQuadtreeNode();
 
 public:
-	long					x0, y0, x1, y1;
+	int32_t					x0, y0, x1, y1;
 	CTerrainQuadtreeNode *	NW_Node;
 	CTerrainQuadtreeNode *	NE_Node;
 	CTerrainQuadtreeNode *	SW_Node;
 	CTerrainQuadtreeNode *	SE_Node;
-	long					Size;
-	long					PatchNum;
+	int32_t					Size;
+	int32_t					PatchNum;
 	D3DXVECTOR3				center;
 	float					radius;
-	BYTE					m_byLODLevel;
+	uint8_t					m_byLODLevel;
 };
